Draft 09 meaning-neutral edits/refactor
* I moved the "Initiation and Discovery" section after "Data Formats"
and "Communication Types" in order to keep all of the protocol sections
together.

* I pulled all of the information on generating signature out of the
Associations section into it's own section.  I also pulled
information about the different signature algorithms into this section.

* In the Generating Signatures section, I refactored the descriptions 
of
Sign-All and Signed-List with an eye toward highlighting the 
differences
between the two.

* I moved Verifying Assertions to it's own section, and made the order
of subsections match the order of the list


git-svn-id: http://svn.openid.net/repos/specifications/authentication/2.0/trunk@23 9099bb4f-571c-0410-8e08-fb6e72fd6b3d