Claimed Identifier patch.
a) Renamed disco_id to claimed_id.

b) Reworded references to the claimed identifier from:

    on which discovery was performed
to:
    which has been discovered

(this is to address the fact the the claimed_id = canonical id for XRIs, in which case the first statement is not accurate)

c) Added a new identifier type - "User-supplied Identifier". This was needed in order to be able to properly define the claimed identifier.
		


git-svn-id: http://svn.openid.net/repos/specifications/authentication/2.0/trunk@109 9099bb4f-571c-0410-8e08-fb6e72fd6b3d